# docs/swagger.yaml
basePath: /api
definitions:
  attendance.CheckInRequest:
    properties:
      employee_id:
        type: string
      latitude:
        type: number
      longitude:
        type: number
    required:
    - employee_id
    - latitude
    - longitude
    type: object
  attendance.CheckOutRequest:
    properties:
      employee_id:
        type: string
      latitude:
        type: number
      longitude:
        type: number
    required:
    - employee_id
    - latitude
    - longitude
    type: object
  auth.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  auth.RegisterRequest:
    properties:
      email:
        type: string
      name:
        maxLength: 100
        minLength: 3
        type: string
      password:
        minLength: 6
        type: string
      role:
        type: string
    required:
    - email
    - name
    - password
    type: object
  auth.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      success:
        type: boolean
    type: object
  common.APIResponse:
    properties:
      data: {}
      error: {}
      message:
        type: string
      success:
        type: boolean
    type: object
  department.CreateDepartmentRequest:
    properties:
      description:
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
      status:
        type: string
    required:
    - name
    type: object
  department.UpdateDepartmentRequest:
    properties:
      description:
        type: string
      name:
        type: string
      status:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: ravi@example.com
    name: Raviraj Sahu
  description: HRMS Backend API using Go, Gin, GORM & PostgreSQL.
  license:
    name: MIT
  termsOfService: http://swagger.io/terms/
  title: HRMS API
  version: "1.0"
paths:
  /attendance:
    get:
      description: Get all attendance records
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.APIResponse'
      security:
      - BearerAuth: []
      summary: Get All Attendance
      tags:
      - Attendance
  /attendance/{id}:
    delete:
      description: Delete attendance record
      parameters:
      - description: Attendance ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete Attendance
      tags:
      - Attendance
    get:
      description: Get attendance by ID
      parameters:
      - description: Attendance ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.APIResponse'
      security:
      - BearerAuth: []
      summary: Get Attendance By ID
      tags:
      - Attendance
  /attendance/check-in:
    post:
      consumes:
      - application/json
      description: Employee check in
      parameters:
      - description: Check In
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/attendance.CheckInRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/common.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.APIResponse'
      security:
      - BearerAuth: []
      summary: Check In
      tags:
      - Attendance
  /attendance/check-out:
    put:
      consumes:
      - application/json
      description: Employee check out
      parameters:
      - description: Check Out
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/attendance.CheckOutRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.APIResponse'
      security:
      - BearerAuth: []
      summary: Check Out
      tags:
      - Attendance
  /attendance/employee/{employee_id}:
    get:
      description: Get attendance by employee
      parameters:
      - description: Employee ID
        in: path
        name: employee_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.APIResponse'
      security:
      - BearerAuth: []
      summary: Get Attendance By Employee
      tags:
      - Attendance
  /departments:
    get:
      description: Get All Departments
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.APIResponse'
      security:
      - BearerAuth: []
      summary: Get All Departments
      tags:
      - Department
    post:
      consumes:
      - application/json
      description: Create a new Department
      parameters:
      - description: Department
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/department.CreateDepartmentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/common.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.APIResponse'
      security:
      - BearerAuth: []
      summary: Create Department
      tags:
      - Department
  /departments/{id}:
    delete:
      description: Delete Department by ID
      parameters:
      - description: Department ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete Department
      tags:
      - Department
    get:
      description: Get All department by ID
      parameters:
      - description: Department ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.APIResponse'
      security:
      - BearerAuth: []
      summary: Get All Department By ID
      tags:
      - Department
    put:
      consumes:
      - application/json
      description: Update Department Details
      parameters:
      - description: Department ID
        in: path
        name: id
        required: true
        type: string
      - description: Department
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/department.UpdateDepartmentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.APIResponse'
      security:
      - BearerAuth: []
      summary: Update Department
      tags:
      - Department
...
  /login:
    post:
      consumes:
      - application/json
      description: Login user
      parameters:
      - description: Login Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/auth.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/auth.Response'
      summary: Login
      tags:
      - Authentication
...
  /profile:
    get:
      description: Get logged-in user profile
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/auth.Response'
      security:
      - BearerAuth: []
      summary: Get Profile
      tags:
      - Authentication
  /register:
    post:
      consumes:
      - application/json
      description: Register a new user
      parameters:
      - description: Register Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/auth.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/auth.Response'
      summary: Register User
      tags:
      - Authentication
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"